Add Serial Port Support to packer-plugin-nutanix (#226)
* Working Config

Working module that will add 1 or more serial ports to a new Nutanix VM when SerialPorts=1.

* Changed SerialPort from int to bool

Changed Serial Port variable from int to bool.  This limits a new VM to just a single serial port, which is all that is needed for this use case.  While there are potential use cases for multiple serial ports on a VM, I'm not sure if they're really needed in most cases because serial ports should usually be removed because they are not used or needed.

* Documentation Update

Add Serialport to builder documentation with description of what it does.

* Update driver.go

Remove SerialIndex++ vestigial code.