fix: Reject subnet name that matches more than one subnet (#248)
* fix: Reject subnet name that matches more than one subnet

It is possible to create multiple overlay subnets with the same name,
and multiple VLAN subnets with the same name within one Prism Element.

Currently, when the plugin finds multiple subnets matching the name, it
select the first match. The chosen subnet may not work, and Packer will
fail, but the user will have no idea why.

Instead, this PR returns an error when the plugin finds multiple subnets
matching the name.

* fixup! fix: Reject subnet name that matches more than one subnet

Try to narrow down to one subnet by filtering by the cluster UUID